+++ Joker.Kafka/KSql/Query/KSqlQueryGenerator.cs
@@ -1,9 +1,11 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Linq;
 using System.Linq.Expressions;
 using Kafka.DotNet.ksqlDB.Infrastructure.Extensions;
 using Kafka.DotNet.ksqlDB.KSql.Linq;
 using Kafka.DotNet.ksqlDB.KSql.Query.Context;
+using Kafka.DotNet.ksqlDB.KSql.Query.Visitors;
 using Kafka.DotNet.ksqlDB.KSql.Query.Windows;
 using Pluralize.NET;
 
@@ -27,17 +29,32 @@ namespace Kafka.DotNet.ksqlDB.KSql.Query
     {
       kSqlVisitor = new KSqlVisitor();
       whereClauses = new Queue<Expression>();
+      joinTables = new List<IEnumerable<Expression>>();
 
       Visit(expression);
+      
+      string finalStreamName = queryContext.StreamName ?? InterceptStreamName(streamName);
 
-      kSqlVisitor.Append("SELECT ");
+      if (joinTables.Any())
+      {
+        var joinsVisitor = new KSqlJoinsVisitor(kSqlVisitor.StringBuilder, options, new QueryContext { StreamName = finalStreamName });
 
-      if (@select != null)
-        kSqlVisitor.Visit(@select.Body);
+        foreach (var joinTable in joinTables)
+        {
+          joinsVisitor.VisitJoinTable(joinTable.ToArray());
+        }
+      }
       else
-        kSqlVisitor.Append("*");
+      {
+        kSqlVisitor.Append("SELECT ");
 
-      kSqlVisitor.Append($" FROM {queryContext.StreamName ?? InterceptStreamName(streamName)}");
+        if (@select != null)
+          kSqlVisitor.Visit(@select.Body);
+        else
+          kSqlVisitor.Append("*");
+
+        kSqlVisitor.Append($" FROM {finalStreamName}");
+      }
 
       bool isFirst = true;
 
@@ -208,6 +225,15 @@ namespace Kafka.DotNet.ksqlDB.KSql.Query
         VisitChained(methodCallExpression);
       }
 
+      if (methodInfo.Name == nameof(QbservableExtensions.Join))
+      {
+        var joinTable = methodCallExpression.Arguments.Skip(1);
+        
+        joinTables.Add(joinTable);
+
+        VisitChained(methodCallExpression);
+      }
+
       return methodCallExpression;
     }
 
@@ -224,6 +250,7 @@ namespace Kafka.DotNet.ksqlDB.KSql.Query
     private TimeWindows windowedBy;
     private LambdaExpression groupBy;
     private LambdaExpression having;
+    private List<IEnumerable<Expression>> joinTables;
 
     protected static Expression StripQuotes(Expression expression)
     {